+++ projectBlueWater/src/main/java/com/lasthopesoftware/bluewater/client/playback/file/MediaPlayerPlaybackHandler.java
@@ -65,7 +65,7 @@ implements
 	}
 
 	@Override
-	public int getCurrentPosition() {
+	public long getCurrentPosition() {
 		try {
 			return isPlaying()
 				? previousMediaPlayerPosition = mediaPlayer.getCurrentPosition()
@@ -77,7 +77,7 @@ implements
 	}
 
 	@Override
-	public int getDuration() {
+	public long getDuration() {
 		try {
 			return mediaPlayer.getDuration();
 		} catch (IllegalStateException e) {
